refactor(game): extract RoundManager delegation properties into RoundManagerDelegationMixin (#1271)
Pulls the read-only / setter @property facade that keeps the round-scoped
public interface identical (round, total_rounds, deadline, current_song,
last_round, round_start_time, round_duration, song_stopped, round_analytics,
the intro-mode flags, early_reveal, songs_remaining, metadata_pending) out of
the state.py God-Object into a new RoundManagerDelegationMixin. Pure code-move,
byte-identical API: the properties stay pass-through delegators to the owned
RoundManager (and PlaylistManager for songs_remaining). The _set_phase
chokepoint and all round-end SSOT stay on GameState.

state.py: 898 -> 783 lines (under the 800 target).
